Dimentia-AI/requirements.txt
openai-whisper
ffmpeg-python
soundfile
miniaudio
pandas
tqdm
numpy
scipy
scikit-learn
joblib
nltk
